Move JsonLike into optic-domain
