+++ workspaces/optic-domain/src/index.ts
@@ -1,10 +1,10 @@
 export * as AsyncTools from './async-tools';
 export * as Streams from './streams';
 export * as Commands from './commands';
-export * as Events from './events';
+export * as Shapes from './shapes';
 export * as Interactions from './interactions';
 
 export * from './commands';
-export * from './events';
+export * from './shapes';
 export * from './interactions';
 export * from './streams';
