+++ workspaces/ui-v2/src/components/ShapeRenderer/ShapeRowBase.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import makeStyles from '@material-ui/styles/makeStyles';
+import { JsonLike } from '@useoptic/optic-domain';
+
 import { IndentSpaces, useSharedStyles } from './SharedStyles';
-import { IFieldRenderer, IShapeRenderer, JsonLike } from '<src>/types';
+import { IFieldRenderer, IShapeRenderer } from '<src>/types';
 import { ShapePrimitiveRender, UnknownPrimitiveRender } from './ShapePrimitive';
 import { useDepth } from './DepthContext';
 import classNames from 'classnames';
